requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
# created-by: conda 24.11.2
annotated-types=0.6.0=py310hca03da5_0
antlr-python-runtime=4.9.3=pyhd8ed1ab_1
appdirs=1.4.4=pyhd3eb1b0_0
appnope=0.1.2=py310hca03da5_1001
asttokens=2.0.5=pyhd3eb1b0_0
blas=1.0=openblas
brotli=1.0.9=h80987f9_9
brotli-bin=1.0.9=h80987f9_9
brotli-python=1.0.9=py310h313beb8_8
bzip2=1.0.8=h80987f9_6
ca-certificates=2024.12.31=hca03da5_0
certifi=2024.12.14=py310hca03da5_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
click=8.1.7=py310hca03da5_0
colorama=0.4.6=py310hca03da5_0
comm=0.2.1=py310hca03da5_0
contourpy=1.3.1=py310h48ca7d4_0
cpython=3.10.16=py310hd8ed1ab_1
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.8.11=py310h313beb8_0
decorator=5.1.1=pyhd3eb1b0_0
docker-pycreds=0.4.0=pyhd3eb1b0_0
eval_type_backport=0.2.2=pyha770c72_0
exceptiongroup=1.2.0=py310hca03da5_0
executing=0.8.3=pyhd3eb1b0_0
filelock=3.13.1=py310hca03da5_0
fonttools=4.51.0=py310h80987f9_0
freetype=2.12.1=hadb7bae_2
giflib=5.2.2=h80987f9_0
gitdb=4.0.7=pyhd3eb1b0_0
gitpython=3.1.43=py310hca03da5_0
gmp=6.2.1=hc377ac9_3
gmpy2=2.1.2=py310h8c48613_0
graphql-core=1.1=py_1
hydra-core=1.3.2=pyhd8ed1ab_0
idna=3.7=py310hca03da5_0
importlib_resources=6.4.0=py310hca03da5_0
ipykernel=6.29.5=py310hca03da5_0
ipython=8.30.0=py310hca03da5_0
jedi=0.19.2=py310hca03da5_0
jinja2=3.1.4=py310hca03da5_1
joblib=1.4.2=py310hca03da5_0
jpeg=9e=h80987f9_3
jupyter_client=8.6.0=py310hca03da5_0
jupyter_core=5.7.2=py310hca03da5_0
kiwisolver=1.4.4=py310h313beb8_0
lcms2=2.15=h481adae_0
lerc=4.0.0=h313beb8_0
libbrotlicommon=1.0.9=h80987f9_9
libbrotlidec=1.0.9=h80987f9_9
libbrotlienc=1.0.9=h80987f9_9
libcxx=14.0.6=h848a8c0_0
libdeflate=1.17=h80987f9_1
libffi=3.4.4=hca03da5_1
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libjpeg-turbo=2.0.0=h1a28f6b_0
liblzma=5.6.3=h39f12f2_1
libopenblas=0.3.21=h269037a_0
libpng=1.6.44=hc14010f_0
libprotobuf=3.20.3=hb5ab8b9_0
libsodium=1.0.18=h1a28f6b_0
libsqlite=3.47.2=h3f77e49_0
libtiff=4.5.0=h5dffbdd_2
libwebp=1.3.2=ha3663a8_0
libwebp-base=1.3.2=h80987f9_1
libxcb=1.13=h9b22ae9_1004
libzlib=1.3.1=h8359307_2
lightning-utilities=0.11.9=py310hca03da5_0
llvm-openmp=14.0.6=hc6e5704_0
lz4-c=1.9.4=h313beb8_1
markupsafe=2.1.3=py310h80987f9_0
matplotlib=3.10.0=py310hb6292c7_0
matplotlib-base=3.10.0=py310h7ef442a_0
matplotlib-inline=0.1.6=py310hca03da5_0
mpc=1.1.0=h8c48613_1
mpfr=4.0.2=h695f6f0_1
mpmath=1.3.0=py310hca03da5_0
ncurses=6.5=h7bae524_1
nest-asyncio=1.6.0=py310hca03da5_0
networkx=3.2.1=py310hca03da5_0
numpy=2.0.1=py310h3b2db8e_1
numpy-base=2.0.1=py310ha9811e2_1
omegaconf=2.3.0=pyhd8ed1ab_0
openjpeg=2.5.2=h54b8e55_0
openssl=3.4.0=h39f12f2_0
packaging=24.2=py310hca03da5_0
parso=0.8.4=py310hca03da5_0
pexpect=4.8.0=pyhd3eb1b0_3
pillow=9.4.0=py310h5a7539a_1
pip=24.2=py310hca03da5_0
platformdirs=3.10.0=py310hca03da5_0
promise=2.3=py310hca03da5_0
prompt-toolkit=3.0.43=py310hca03da5_0
prompt_toolkit=3.0.43=hd3eb1b0_0
protobuf=3.20.3=py310h313beb8_0
psutil=5.9.0=py310h1a28f6b_0
pthread-stubs=0.3=h1a28f6b_1
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pydantic=2.10.3=py310hca03da5_0
pydantic-core=2.27.1=py310h2aea54e_0
pygments=2.15.1=py310hca03da5_1
pyparsing=3.2.0=py310hca03da5_0
pysocks=1.7.1=py310hca03da5_0
python=3.10.16=h870587a_1_cpython
python-dateutil=2.9.0post0=py310hca03da5_2
python_abi=3.10=2_cp310
pytorch=2.5.1=py3.10_0
pyyaml=6.0.2=py310h80987f9_0
pyzmq=26.2.0=py310h313beb8_0
readline=8.2=h1a28f6b_0
requests=2.32.3=py310hca03da5_1
scikit-learn=1.5.2=py310h313beb8_0
scipy=1.14.1=py310hd336fd7_0
sentry-sdk=2.18.0=py310hca03da5_0
setproctitle=1.2.2=py310h80987f9_1
setuptools=75.1.0=py310hca03da5_0
six=1.16.0=pyhd3eb1b0_1
smmap=4.0.0=pyhd3eb1b0_0
sqlite=3.47.2=hd7222ec_0
stack_data=0.2.0=pyhd3eb1b0_0
sympy=1.13.3=pyh2585a3b_104
threadpoolctl=3.5.0=py310h33ce5c2_0
tk=8.6.13=h5083fa2_1
torchaudio=2.5.1=py310_cpu
torchmetrics=1.6.1=pyhd8ed1ab_0
torchvision=0.20.1=py310_cpu
tornado=6.4.2=py310h80987f9_0
tqdm=4.67.1=pyhd8ed1ab_1
traitlets=5.14.3=py310hca03da5_0
typing=3.10.0.0=py310hca03da5_0
typing-extensions=4.12.2=py310hca03da5_0
typing_extensions=4.12.2=py310hca03da5_0
tzdata=2024b=h04d1e81_0
unicodedata2=15.1.0=py310h80987f9_0
urllib3=2.2.3=py310hca03da5_0
wandb=0.19.2=py310hde4708a_0
wcwidth=0.2.5=pyhd3eb1b0_0
wheel=0.44.0=py310hca03da5_0
xorg-libxau=1.0.12=h5505292_0
xorg-libxdmcp=1.1.5=hd74edd7_0
xz=5.4.6=h80987f9_1
yaml=0.2.5=h1a28f6b_0
zeromq=4.3.5=h313beb8_0
zlib=1.3.1=h8359307_2
zstd=1.5.6=hb46c0d2_0
